task2

+++ calculator/src/main/java/com/example/calculator/service/CalculatorService.java
@@ -1,17 +1,11 @@
 package com.example.calculator.service;
 
-import org.springframework.stereotype.Repository;
 import org.springframework.stereotype.Service;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.ResponseStatus;
-
-import java.util.List;
 
 @Service
 public class CalculatorService {
 
-    public String calculate(Integer value1, Integer value2, String ops) {
+    public String calculate(Double value1, Double value2, String ops) {
         switch (ops) {
             case "add":
                 return value1 + " + " + value2 + " is " + (value1 + value2);
